Move chapter 5 listings to standard file layout.
